edge-gateway/requirements.txt
# SmartWatts Edge Gateway Requirements
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# MQTT Communication
paho-mqtt==1.6.1
asyncio-mqtt==0.16.1

# Modbus Integration
pymodbus==3.5.2
pyserial==3.5

# Database
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0

# AI/ML
tensorflow-lite==2.14.0
numpy==1.24.3
scikit-learn==1.3.2

# Data Processing
pandas==2.1.4
pytz==2023.3

# Network & Communication
httpx==0.25.2
websockets==12.0

# Device Management
psutil==5.9.6
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
